+++ 11-flagship/01-opslane/internal/handlers/handlers.go
@@ -1,6 +1,5 @@
 // Copyright (c) 2026 Rasel Hossen
-// Licensed under The Go Engineer License v1.0
-// Commercial use is prohibited without permission.
+// See LICENSE for usage terms.
 
 package handlers
 
@@ -12,9 +11,9 @@ import (
 	"log/slog"
 	"net/http"
 
-	"github.com/rasel9t6/the-go-engineer/11-flagship/01-enterprise-capstone/internal/middleware"
-	"github.com/rasel9t6/the-go-engineer/11-flagship/01-enterprise-capstone/internal/models"
-	"github.com/rasel9t6/the-go-engineer/11-flagship/01-enterprise-capstone/internal/repository"
+	"github.com/rasel9t6/the-go-engineer/11-flagship/01-opslane/internal/middleware"
+	"github.com/rasel9t6/the-go-engineer/11-flagship/01-opslane/internal/models"
+	"github.com/rasel9t6/the-go-engineer/11-flagship/01-opslane/internal/repository"
 	"golang.org/x/crypto/bcrypt"
 )